Added functionality to some scripts
OverviewScreen.cs
-Created a system that will instantiate game panels in a grid using the Grid Layout Group in the Category Panels and Question Panels GameObjects.

GamePanel.cs
-Getting the game mode now relies on the Game Manager instead of a dedicated game mode variable.

+++ Assets/scripts/GameManager.cs
@@ -19,9 +19,9 @@ public class GameManager : MonoBehaviour
     [Range(1, 3)]
     public int teamCount;
 
-    public float teamOneScore;
-    public float teamTwoScore;
-    public float teamThreeScore;
+    public int teamOneScore;
+    public int teamTwoScore;
+    public int teamThreeScore;
 
 
     private void Awake()
